requirements.txt
# Hand Gesture Monitoring System
# Python 3.9+

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Video & hand tracking
opencv-python>=4.8.0
mediapipe>=0.10.0

# ML
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.13.0

# Utilities
python-dotenv>=1.0.0
werkzeug>=2.3.0

# Optional: for plots during training
matplotlib>=3.7.0
